requirements.txt
numpy==1.18.5
environs==8.0.0
PyYAML==5.3.1
matplotlib==3.2.2
